requirements.txt
pdfminer.six
nltk
sentencepiece
torch
numpy
